// src/app/models/customer.model.ts
export interface Customer {
    id: any;
    first_name: string;
    last_name: string;
    cnic: string;
    mobile: string;
    email: string;
    password: string;
    // address: string;
    // package: any;
    // routerOf: any;
    // routerBrand: any;
    // routerModel: any;
    // routerPrice: any;
    // dropWireOf: any;
    // dropWireLength: any;
    // dropWirePricePL: any;
}